Create Principal Component Analysis(Dimension Reduction)
During EDA when we have a large set of variables we often get confused and lost deciding which features to choose so that our model will be safe from overfitting. In such cases, Dimension Reduction comes into play. In these cases, we can make use of an unsupervised Algorithm known as PCA. PCA is used to find inter-relation between variables in the data y reducing the dimensions of a feature in such a way that they are statistically independent & not correlated.